got NAMD from test repo
I'd created NAMD in a test repo.  This is just a copy of that file.  This was done before the current full instructions and scrubber scripts were tested, so it's possible there are a couple of niggling unprintable characters.  This should be checked more carefully than the other pages as a result before it's published.